requirements.txt
aiogram>=3.0.0
web3>=6.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
eth-account>=0.8.0
